+++ cookbook/samples/wip/README.md
@@ -1,0 +1,5 @@
+# WIP
+
+## Test the new execute_mission API
+
+- it tracks costs related to a specific mission_id
