+++ pyproject.toml
@@ -26,14 +26,7 @@ include = "cookbook"
 
 [tool.poetry.dependencies]
 beautifulsoup4 = "^4.13.4"
-# pipelex = { version = "0.2.0", extras = [
-#     "mistralai",
-#     "anthropic",
-#     "google",
-#     "bedrock",
-#     "fal",
-# ] }
-pipelex = { path = "../pipelex", develop = true, extras = [
+pipelex = { git = "https://github.com/Pipelex/pipelex.git", branch = "feature/Mission-tracker", extras = [
     "mistralai",
     "anthropic",
     "google",
